requirements.txt
iso8601
Django
south
sentry
django-taggit
django-haystack
